+++ lib/cftp.c
@@ -272,7 +272,7 @@ int cftp_client_receive_file(struct cftp_descriptor *desc, const char *file_in, 
 					print_char('.');
 				}
 
-				if (recvlen < max_xfer_length)
+				if ((size_t)recvlen < max_xfer_length)
 				{
 					println(" Receive data complete");
 					cftp_send_ack_message(desc, (struct cftp_ack_message *)msg, blk_num, 0);
@@ -534,7 +534,7 @@ int cftp_client_send_file(struct cftp_descriptor *desc, const char *file_in, u32
 		case CFTP_PACKAGE_ACK:
 			if (msg->ack_msg.blk_num == blk_num)
 			{
-				if (readlen < max_data_length)
+				if ((size_t)readlen < max_data_length)
 				{
 					ret = 0;
 					progress_bar_finish(&bar);
@@ -679,7 +679,7 @@ int cftp_server_receive_file(struct cftp_descriptor *desc, const char *filename,
 
 			blk_num++;
 
-			if (recvlen < max_xfer_length)
+			if ((size_t)recvlen < max_xfer_length)
 			{
 				cftp_send_ack_message(desc, (struct cftp_ack_message *)msg, blk_num, 0);
 				progress_bar_finish(&bar);
@@ -826,7 +826,7 @@ label_send_data:
 				goto label_send_data;
 			}
 
-			if (readlen < max_data_length)
+			if ((size_t)readlen < max_data_length)
 			{
 				progress_bar_finish(&bar);
 				println("Send data complete");
